Refactor and addition of Twitch data route
